+++ core/src/commonMain/kotlin/com/deezer/caupain/formatting/json/JsonFormatter.kt
@@ -24,31 +24,24 @@
 
 package com.deezer.caupain.formatting.json
 
-import com.deezer.caupain.formatting.FileFormatter
+import com.deezer.caupain.formatting.SinkFormatter
 import com.deezer.caupain.formatting.model.Input
-import com.deezer.caupain.internal.DefaultFileSystem
 import com.deezer.caupain.internal.IODispatcher
 import com.deezer.caupain.serialization.DefaultJson
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.json.okio.encodeToBufferedSink
 import okio.BufferedSink
-import okio.FileSystem
-import okio.Path
 
 /**
- * JsonFormatter is a [FileFormatter] that formats the output as JSON.
+ * JsonFormatter is a [SinkFormatter] that formats the output as JSON.
  *
- * @param path The path to the JSON file to write.
- * @param fileSystem The file system to use for writing the file. Default uses the native file system.
  * @param ioDispatcher The coroutine dispatcher to use for IO operations. Default uses IO dispatcher.
  */
 @OptIn(ExperimentalSerializationApi::class)
 public class JsonFormatter(
-    path: Path,
-    fileSystem: FileSystem = DefaultFileSystem,
     ioDispatcher: CoroutineDispatcher = IODispatcher,
-) : FileFormatter(path, fileSystem, ioDispatcher) {
+) : SinkFormatter(ioDispatcher) {
 
     override suspend fun BufferedSink.writeUpdates(input: Input) {
         DefaultJson.encodeToBufferedSink(input, this)