+++ core/src/commonTest/kotlin/com/deezer/caupain/formatting/MarkdownFormatterTest.kt
@@ -27,8 +27,6 @@ package com.deezer.caupain.formatting
 import com.deezer.caupain.formatting.markdown.MarkdownFormatter
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import okio.FileSystem
-import okio.Path
 import org.intellij.lang.annotations.Language
 
 @OptIn(ExperimentalCoroutinesApi::class)
@@ -42,11 +40,8 @@ class MarkdownFormatterTest : FileFormatterTest() {
     override val fullResult: String
         get() = FULL_RESULT
 
-    override fun createFormatter(
-        fileSystem: FileSystem,
-        path: Path,
-        ioDispatcher: CoroutineDispatcher
-    ): FileFormatter = MarkdownFormatter(path, fileSystem, ioDispatcher)
+    override fun createFormatter(ioDispatcher: CoroutineDispatcher) =
+        MarkdownFormatter(ioDispatcher)
 }
 
 @Language("Markdown")